benchmarks: extract append_filter_by_tier to tiers.py
So that we do not duplicate it in aggregate.py.
